WIP: Add a comparator for bigsets
See the BSComparator for the spec. The purpose is to avoid sext/whatever
encoding and decoding bigset data when reading a writing. Instead use an
erlang binary, and have the comparator pick it a part to sort sets so
that sets are together, clocks come first, elements are ordered etc
etc. This is all copied from Engel's work. The mistakes of course are
mine. I am sure this is horrible c++, but it is a prototype!